requirements.txt
flask-cors
werkzeug
flask-mongoengine
bcrypt
pyjwt
dnspython
jinja2
openpyxl
docxtpl
slackclient
flask-admin
email_validator
pptx_template_simple
